+++ CHANGELOG.md
@@ -24,6 +24,9 @@
 - Allow changing window transparency style in settings.
 - Show scrollbar when tree overflows at x axis.
 - Resizable panels. Drag at borders to change panel size.
+- Global searcher, press `ctrl+p` to open it.
+  - Search assets using their ids or names.
+  - Search tags using their ids or names.
 
 # What's Fixed